adds cedar-loader service to lazy-load dependencies
